requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.61.2
h11==0.11.0
pydantic==1.7.2
python-magic==0.4.18
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
uvicorn==0.12.3
